+++ cg_algo.py
@@ -3,6 +3,36 @@ from OpenGL.GLUT import *
 from OpenGL.GLU import *
 from math import *
 
+
+def drawPoints(xc, yc, x, y):
+    glVertex2i(xc + x, yc + y)
+    glVertex2i(xc - x, yc + y)
+    glVertex2i(xc + x, yc - y)
+    glVertex2i(xc - x, yc - y)
+    glVertex2i(xc + y, yc + x)
+    glVertex2i(xc - y, yc + x)
+    glVertex2i(xc + y, yc - x)
+    glVertex2i(xc - y, yc - x)
+
+
+def mid_point_circle(center, radius):
+    x = 0
+    y = radius
+    p = 1 - radius
+    glBegin(GL_POINTS)
+    drawPoints(center[0],center[1],x,y)
+    while x <= y:
+        x += 1
+        if p < 0:
+            p = p + 2 * x + 1
+        else:
+            y -= 1
+            p = p + 2 * (x - y) + 1
+        drawPoints(center[0],center[1],x,y)
+    glEnd()
+    glFlush()
+
+
 def drawLineLow(x1,y1,x2,y2):
     dx = x2 - x1
     dy = y2 - y1
@@ -27,6 +57,7 @@ def drawLineLow(x1,y1,x2,y2):
         glVertex2i(x,y)
     glEnd()
     glFlush()
+
 
 def drawLineHigh(x1,y1,x2,y2):
     dx = x2 - x1
@@ -53,7 +84,8 @@ def drawLineHigh(x1,y1,x2,y2):
     glEnd()
     glFlush()
 
-def BresenhamLine(start, dest):
+
+def bresenham_line(start, dest):
     if(abs(dest[1]-start[1]) == 0):
         x = min(start[0], dest[0])
         glBegin(GL_POINTS)
